% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_log_handlers.R
\name{vt_scrape_change_log}
\alias{vt_scrape_change_log}
\alias{vt_kable_change_log}
\alias{vt_use_change_log}
\title{Scrape change log from a validation project}
\usage{
vt_scrape_change_log()

vt_kable_change_log(change_log_info, format = vt_render_to())

vt_use_change_log(date = NULL, version = NULL, open = interactive())
}
\arguments{
\item{change_log_info}{data.frame as exported from \code{\link{vt_scrape_change_log}}}

\item{format}{passed to \code{knitr::kable}}

\item{date}{passed to template}

\item{version}{version to set in news file}

\item{open}{whether to open the file after}
}
\value{
data.frame with variables \code{version}, \code{effective_date}, \code{description}

a knitr_kable object

path to change log file, used for side effect of creating change_log
}
\description{
Scrape change log from a validation project

Format change log info table for validation report

Initiate a change_log file
}
\note{
Extracts validation version, date, and description from change log items
that start with \code{[validation]}.
}
\examples{

withr::with_tempdir({
 file.create(".here")
 vt_use_validation()

 vt_use_change_log()

 log_data <- vt_scrape_change_log()
 print(log_data)

 vt_kable_change_log(log_data)

})
}
