% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{vt_use_validation_config}
\alias{vt_use_validation_config}
\alias{vt_add_user_to_config}
\title{use a validation config file}
\usage{
vt_use_validation_config(
  pkg = ".",
  val_dir = "vignettes/validation",
  val_dir_o = "validation",
  username_list = list(),
  ...
)

vt_add_user_to_config(username = whoami::username(), name, title, pkg = ".")
}
\arguments{
\item{val_dir}{\link{character} where validation contents are used interactively}

\item{val_dir_o}{\link{character} which folder under \code{inst} should the contents for
validation be copied to on successful validation build.}

\item{username_list}{named list of named vectors. Name of list is username,
with named list containing entries for username, name, and
title to be used for documentation.}

\item{...}{These dots are for future extensions and must be empty.}

\item{username}{computer username associated with the name and role}

\item{name}{name of the user}

\item{path}{where to write config file}

\item{role}{title of the user}
}
\value{
Used for side effect to create validation config file. Invisibly
returns TRUE on success.

Used for side effect of adding user information to validation config
file. Invisibly returns TRUE on success.
}
\description{
validation configuration for directories and username-name-role key

Add user information to the projects validation config file to make for easier documentation
}
