% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_news.R
\name{convert_allnews}
\alias{convert_allnews}
\title{Convert NEWS to data.frame for all release versions listed}
\usage{
convert_allnews(news_log)
}
\arguments{
\item{news_log}{vector of NEWS, using \code{read_lines}}
}
\value{
a data.frame with variables: Version, Comment, Date. one row for all
validation comments. Other comment flags ignored
}
\description{
Convert NEWS to data.frame for all release versions listed
}
\section{Last Updated}{

2020-05-15
}

\author{
Marie Vendettuoli
}
\keyword{internal}
