% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecy_vers.R
\name{dep_versions}
\alias{dep_versions}
\title{Dependency versions}
\usage{
dep_versions()
}
\value{
vector of package versions for the list in  DESCRIPTION imports from current project
}
\description{
Retrieves the DESCRIPTION Imports field from current project and
identifies what versions are installed on current system.
}
