% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_function_authors.R
\name{.scrape_one_function_file}
\alias{.scrape_one_function_file}
\title{Retrieve function author for one .R file}
\usage{
.scrape_one_function_file(one_file)
}
\description{
Retrieve function author for one .R file
}
\keyword{internal}
