% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_referencing.R
\name{dynamic_reference_rendering}
\alias{dynamic_reference_rendering}
\title{Dynamic Reference Rendering}
\usage{
dynamic_reference_rendering(input, reference = NULL)
}
\arguments{
\item{input}{R object or path to the file to convert dynamic referencing to values}

\item{reference}{which dynamic referencer to use. When NULL, uses internal dynamic referencer.}
}
\value{
text with dynamic referencing evaluated
}
\description{
enable dynamic referencing by reading file and converting any dynamic references
into their values for rendering in the validation report.
}
