% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_spec_roxy_blocks.R
\name{eval_processing_tc_code}
\alias{eval_processing_tc_code}
\title{eval_processing_tc_code}
\usage{
eval_processing_tc_code(locations, lines, envir)
}
\arguments{
\item{locations}{postion in \code{lines} which correspond to start and stop
of code chunks}

\item{lines}{RMD source with previously rendered sections omitted.}

\item{envir}{Environment to evaluate R code chunks. Allows for later chunks
in single RMD to depend on earlier ones.}
}
\description{
Evaluate the procesing test code from .RMD file when the top
section has been identified as a code chunk (should start at first line)
}
