% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_test_code.R
\name{eval_test_code}
\alias{eval_test_code}
\title{Evaluate the test code}
\usage{
eval_test_code(one_file, test_env = new.env())
}
\arguments{
\item{one_file}{full path to test code file.}

\item{test_env}{environment to perform tests in}
}
\value{
a table with variables: \code{Test}, \code{Expected}, \code{Results},
\code{Pass/Fail}. Suitable for including in validation report
}
\description{
Evaluate the test code
}
