% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_func_roxy_blocks.R
\name{get_section_contents}
\alias{get_section_contents}
\title{Retrieve the value in a function specification block of a custom section}
\usage{
get_section_contents(tag, block)
}
\arguments{
\item{tag}{name of the section, case insensitive.}

\item{block}{character vector that holds the documentation block}
}
\value{
section value
}
\description{
Retrieve the value in a function specification block of a custom section
}
