% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R, R/code.R, R/spec.R
\name{vt_use_case}
\alias{vt_use_case}
\alias{vt_use_code}
\alias{vt_use_spec}
\title{Create a validation specification, test case, or test code file}
\usage{
vt_use_case(case_name, pkg = ".")

vt_use_code(code_name, pkg = ".")

vt_use_spec(spec_name, pkg = ".")
}
\arguments{
\item{case_name}{The name/path of the test case. Cases can be named with your
file system separator and will be organized as a directory structure. Cases
are located at \code{./inst/validation/cases/{case_name}}.}

\item{pkg}{Top-level of package}

\item{code_name}{The name/path of the test code. Codes can be named with your
file system separator and will be organized as a directory structure. Codes
are located at \code{./inst/validation/codes/{code_name}}.}

\item{spec_name}{The name/path of the spec. specs can be named with your
file system separator and will be organized as a directory structure. Specs
are located at \code{./inst/validation/specs/{spec_name}}.}
}
\description{
Create a validation specification, test case, or test code file
}
\examples{
package_dir <- tempdir()

vt_use_validation(package_dir)

# Create spec at the cases top level `inst/validation/cases/spec1`
vt_use_case("spec1", package_dir)
# Create spec at `inst/validation/cases/regTests/Update2/spec2`
vt_use_case("regTests/Update2/spec2", package_dir)
}
