% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_spec_roxy_blocks.R
\name{process_single_rmd}
\alias{process_single_rmd}
\title{process_single_rmd}
\usage{
process_single_rmd(lines, envir = new.env())
}
\arguments{
\item{lines}{RMD source with previously rendered sections omitted.}

\item{envir}{environment for evaluating R code. Allows for later chunks in
single RMD to depend on earlier ones.}
}
\description{
Recursive function to handle processing test code RMD files.
}
