% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_table.R
\name{vt_scrape_risk_assessment}
\alias{vt_scrape_risk_assessment}
\alias{vt_kable_risk_assessment}
\title{Scrape "riskAssessment" tag in requirements to generate table}
\usage{
vt_scrape_risk_assessment(reference = NULL, src = ".", ref = vt_path())

vt_kable_risk_assessment(x, format = NULL)
}
\arguments{
\item{reference}{dynamic reference holder if it already exists}

\item{src, ref}{passed to \code{\link{vt_scrape_tags_from}}}

\item{x}{data.frame as output from \code{\link{vt_scrape_risk_assessment}}}

\item{format}{passed to \code{kable}}
}
\value{
a data.frame documenting requirements to risk assessments

knitr_kable object
}
\description{
Scrape "riskAssessment" tag in requirements to generate table

Kable handler for output of \code{\link{vt_scrape_risk_assessment}}
}
