% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_coverage_matrix.R
\name{scrape_coverage_block}
\alias{scrape_coverage_block}
\title{Generate a coverage matrix from the specification reference in test cases}
\usage{
scrape_coverage_block(files)
}
\arguments{
\item{files}{path to folder containing .Rmd file to scrape}
}
\value{
data.frame with variables: \code{test_case}, \code{specifications}, and
\code{value = "x"}.
}
\description{
May be used for specification or test case .Rmd files located
in \code{function_specification} folder.
}
