% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_func_roxy_blocks.R
\name{scrape_function_author_block}
\alias{scrape_function_author_block}
\title{Generate at data.frame from the function roxygen documentation
block.}
\usage{
scrape_function_author_block(one_file)
}
\arguments{
\item{one_file}{full filepath of the .R file to scrape}
}
\value{
data.frame with variables: \code{title}, \code{author}, and
\code{last_updated_date}.
}
\description{
May be used for specification or test case .Rmd files located
in \code{function_specification} folder.
}
