% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_func_roxy_blocks.R
\name{scrape_function_specification_block}
\alias{scrape_function_specification_block}
\title{Generate at data.frame from the function specification roxygen documentation
block.}
\usage{
scrape_function_specification_block(one_file)
}
\arguments{
\item{one_file}{full filepath of the .Rmd file to scrape}
}
\value{
data.frame with variables: \code{title}, \code{last_update_by}, and
\code{last_updated_date}.
}
\description{
May be used for specification or test case .Rmd files located
in \code{function_specification} folder.
}
\examples{
\dontrun{
  library(here)
  library(magrittr)
  list.files(here("vignettes", "function_specifications", "test_cases"),
           full.names = TRUE)[1:4] \%>\%
  lapply(., scrape_function_specification_block) \%>\%
  do.call(rbind, .)

  list.files(here("vignettes", "function_specifications", "specifications"),
            full.names = TRUE)[1:4] \%>\%
  lapply(., scrape_function_specifications) \%>\%
  do.call(rbind, .)
}
}
