% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_news.R
\name{scrape_news}
\alias{scrape_news}
\title{SCrape NEWS file for version validation information}
\usage{
scrape_news(file)
}
\arguments{
\item{file}{path to NEWS markdown file}
}
\value{
a data.frame with variables: Version, Comment, Date. one row for all
validation comments. Other comment flags ignored
}
\description{
SCrape NEWS file for version validation information
}
\section{Last Updated}{

2020-05-15
}

\examples{
\dontrun{
news_log <- read_lines(file.path(here(), "NEWS.md"))
scrape_news(news_log)
}
}
\author{
Marie Vendettuoli
}
