% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_spec_roxy_blocks.R
\name{scrape_processing_rmd}
\alias{scrape_processing_rmd}
\alias{scrape_spec_rmd}
\title{Scrape a processing test code RMD file}
\usage{
scrape_spec_rmd(file, envir = new.env())
}
\arguments{
\item{file}{character vector specifying file location}

\item{envir}{Environment to evaluate R code chunks. Allows for later chunks
in single RMD to depend on earlier ones.}
}
\description{
Scrape a processing test code RMD file
}
