% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_roxygen.R
\name{scrape_roxygen}
\alias{scrape_roxygen}
\title{Scrape Roxygen blocks}
\usage{
scrape_roxygen(file, ..., type = tools::file_ext(file))
}
\arguments{
\item{file}{file to scrape roxygen block from}

\item{...}{These dots are for future extensions and must be empty.}

\item{type}{method of parse_roxygen to use if other that file extension}
}
\value{
a list of roxygen blocks found in the file.
}
\description{
valtools uses roxygen across multiple file types to provide documentation.
this function provides the tooling necessary to scrape from the
major file types that we use ( R, R test code, markdown, Rmarkdown)
and provides a consistent output type to capture the information
necessary to help high level functions make assumptions.
}
