% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_spec_roxy_blocks.R
\name{scrape_test_code_block}
\alias{scrape_test_code_block}
\title{Generate at data.frame from the test code roxygen documentation blocks.}
\usage{
scrape_test_code_block(one_file)
}
\arguments{
\item{one_file}{full filepath of the .R or .r file to scrape}
}
\value{
data.frame with variables: \code{title}, \code{last_update_by}, and
\code{last_updated_date}.
}
\description{
May be used for .R or .r files located
in \code{tests/test_specs} folder.
}
\examples{
\dontrun{
  list.files(here("tests", "spec_tests"), pattern = ".R", ignore.case = TRUE,
           full.names = TRUE)[1:4] \%>\%
    lapply(., scrape_test_code_block) \%>\%
    do.call(rbind, .)
}
}
