% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sig_tab.R
\name{signature_table}
\alias{signature_table}
\title{Generate a signature table}
\usage{
signature_table(people)
}
\arguments{
\item{people}{data frame with variables: role, name, and title}
}
\value{
a kableExtra table with people info formatted for validation report
}
\description{
Generate a signature table
}
