% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{vt_use_validation}
\alias{vt_use_validation}
\alias{vt_create_package}
\title{Create a validation structure}
\usage{
vt_use_validation(pkg = ".")

vt_create_package(pkg = ".", ...)
}
\arguments{
\item{pkg}{Top level directory of a package}

\item{...}{Additional argument passed to \code{usethis::create_package()}}
}
\description{
Creates a structure for validation artifacts. Validation items are stored in
\code{inst/validation}
}
