% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{vt_add_user_to_config}
\alias{vt_add_user_to_config}
\title{Add users to validation config file}
\usage{
vt_add_user_to_config(username = whoami::username(), name, title, pkg = ".")
}
\arguments{
\item{username}{computer username associated with the name and role}

\item{name}{name of the user}

\item{role}{title of the user}
}
\description{
Add user information to the projects validation config file to make for easier documentation
}
