% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_file.R
\name{vt_file}
\alias{vt_file}
\title{print files for report generation}
\usage{
vt_file(file, ..., dynamic_referencing = FALSE)
}
\arguments{
\item{file}{file to evaluate}

\item{...}{These dots are for future extensions and must be empty.}

\item{dynamic_referencing}{Whether to employ dynamic referencing or not.
defaults to FALSE.}
}
\value{
a list of roxygen blocks found in the file.
}
\description{
valtools assists the user in generating the validation report by allowing the
user to define the order in which
}
