% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_table.R
\name{vt_generate_sig_table}
\alias{vt_generate_sig_table}
\title{Generate a signature table for a validation report}
\usage{
vt_generate_sig_table(people)
}
\arguments{
\item{people}{data frame with variables: role, name, and title}
}
\value{
A data frame with people info formatted for validation report with
the
}
\description{
Generate a signature table for a validation report
}
