% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-user.R
\name{vt_get_all_users}
\alias{vt_get_all_users}
\title{Get all users from validation config file without knowing usernames}
\usage{
vt_get_all_users()
}
\arguments{
\item{pkg}{Top-level directory of the package to validate}
}
\value{
list of all users in config file
}
\description{
Get all users from validation config file without knowing usernames
}
