% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_files.R
\name{vt_get_child_files}
\alias{vt_get_child_files}
\title{Identify ordering of validation or user-designated child files}
\usage{
vt_get_child_files(
  loc = c("folder", "yml"),
  validation_order = c("requirements", "test_cases", "test_code")
)
}
\arguments{
\item{loc}{location to explore. Either "folder" for naive inclusion of
validation folder contents, or "yml" to use validation_folder field of validation.yml}

\item{validation_order}{optional ordering of validation folders to search}
}
\value{
vector of child file names to include in validation report
}
\description{
Identify ordering of validation or user-designated child files
}
\examples{
withr::with_tempdir({
  vt_use_validation()
  vt_use_test_case("testcase1", username = "a user", open = FALSE)
  vt_use_req("req1", username = "a user", open = FALSE)
  vt_use_test_code("testcode1", username = "another user", open = FALSE)

  # as listed in validation.yml validation_files
  vt_get_child_files(loc = "yml")

  # as ordered in validation subfolders
  vt_get_child_files(loc = "folder", validation_order = c("requirements", "test_cases", "test_code"))

})
}
