% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage-matrix.R
\name{vt_kable_coverage_matrix}
\alias{vt_kable_coverage_matrix}
\title{Kable handler for output of \code{\link{vt_scrape_coverage_matrix}}}
\usage{
vt_kable_coverage_matrix(x, format = vt_render_to())
}
\arguments{
\item{x}{data.frame as output from \code{\link{vt_scrape_coverage_matrix}}}

\item{format}{passed to \code{kable}}
}
\value{
knitr_kable object
}
\description{
Kable handler for output of \code{\link{vt_scrape_coverage_matrix}}
}
