% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_authors.R
\name{vt_kable_functions}
\alias{vt_kable_functions}
\title{Kable code from function author table}
\usage{
vt_kable_functions(author_details, format = "latex")
}
\arguments{
\item{author_details}{data.frame as exported from \code{vt_scrape_functions}}

\item{format}{passed to \code{knitr::kable}}
}
\value{
knitr_kable object
}
\description{
Kable code from function author table
}
