% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_handlers.R
\name{vt_kable_news}
\alias{vt_kable_news}
\title{Format NEWS info table for validation report}
\usage{
vt_kable_news(news_info, format = "latex")
}
\arguments{
\item{news_info}{data.frame as exported from \code{\link{vt_scrape_news}}}

\item{format}{passed to \code{knitr::kable}}
}
\value{
a knitr_kable object
}
\description{
Format NEWS info table for validation report
}
