% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_specification.R
\name{vt_kable_requirements}
\alias{vt_kable_requirements}
\title{kable code from requirement author table}
\usage{
vt_kable_requirements(requirement_details, format = "latex")
}
\arguments{
\item{requirement_details}{data.frame as exported from vt_scrape_requirements}

\item{format}{passed to \code{knitr::kable}}
}
\description{
kable code from requirement author table
}
