% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_table.R
\name{vt_kable_sig_table}
\alias{vt_kable_sig_table}
\title{Kable defaults for rendering validation report}
\usage{
vt_kable_sig_table(people, format = vt_render_to())
}
\arguments{
\item{people}{A dataframe with the columns role, Name and Title, Signature,
and "Date"}

\item{format}{passed to knitr::kable}
}
\value{
knitr_kable object
}
\description{
Kable defaults for rendering validation report
}
