% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val-env-handlers.R
\name{vt_kable_val_env}
\alias{vt_kable_val_env}
\title{Generates kable code for validation environment details}
\usage{
vt_kable_val_env(val_env, format = "latex")
}
\arguments{
\item{val_env}{data.frame as output from \code{\link{vt_scrape_val_env}}}

\item{format}{passed to \code{knitr::kable}}
}
\value{
knitr_kable object
}
\description{
Generates kable code for validation environment details
}
