% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_table.R
\name{vt_render_sig_table}
\alias{vt_render_sig_table}
\title{Return the templated code for inclusion in the validation report}
\usage{
vt_render_sig_table(template = "sig_table.Rmd", output)
}
\arguments{
\item{template}{Name of the template to render}

\item{output}{directory passed to \code{render_template()}}
}
\value{
The text to include in a validation R markdown document
}
\description{
Return the templated code for inclusion in the validation report
}
