% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_validation_report.R
\name{vt_render_validation_report}
\alias{vt_render_validation_report}
\title{provide a nice wrapper to set states around render}
\usage{
vt_render_validation_report(
  report_path,
  output_dir = dirname(report_path),
  output_file = NULL,
  ...,
  render_time = c("build", "installed"),
  package = ""
)
}
\arguments{
\item{report_path}{path to the validation report rmarkdown}

\item{output_dir}{path to directory to output rendered report. defaults to same folder}

\item{output_file}{expected output filename sans extension}

\item{...}{arguments passed to \link[rmarkdown]{render}}

\item{render_time}{type of rendering of validation to run, "build" or "installed".}

\item{package}{the report  type of rendering of validation to run, "build" or "installed".}
}
\description{
This package is not intended for use by the end users. This is to be used within
valtools packages.
}
