% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_authors.R
\name{vt_scrape_authors}
\alias{vt_scrape_authors}
\title{Scrape author info from R functions}
\usage{
vt_scrape_authors(
  tags = c("editor", "editDate", "export"),
  src = ".",
  ref = vt_path()
)
}
\arguments{
\item{tags}{which tags to keep. defaults to editor, editDate, and export.}

\item{src}{path to package source. defaults to the current directory.
passed to \code{valtools:::scrape_tags_from}}

\item{ref}{reference path to where validation documentation lives. defaults to
vt_path() passed to \code{valtools:::scrape_tags_from}}
}
\description{
Scrape author info from R functions
}
\examples{
\dontrun{
# all functions
vt_scrape_authors(tags = c("editor", "editDate"))

# only public functions
vt_scrape_authors()  \%>\%
  dplyr::filter(!is.na(export)) \%>\%
  dplyr::select(-export)
}
}
