% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage-matrix.R
\name{vt_scrape_coverage_matrix}
\alias{vt_scrape_coverage_matrix}
\title{Scrape "coverage" tag in test code to generate mapping}
\usage{
vt_scrape_coverage_matrix(
  type = c("long", "wide"),
  reference = NULL,
  src = ".",
  ref = vt_path()
)
}
\arguments{
\item{type}{one of "long" or "wide" which determines shape of output table}

\item{reference}{dynamic reference holder if it already exists}

\item{src, ref}{passed to \code{\link{vt_scrape_tags_from}}}
}
\value{
a data.frame mapping requirement ids to test case ids.
}
\description{
Scrape "coverage" tag in test code to generate mapping
}
