% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_function_authors.R
\name{vt_scrape_function_authors}
\alias{vt_scrape_function_authors}
\title{Retrieve authorship details for all functions}
\usage{
vt_scrape_function_authors(pkg = ".")
}
\arguments{
\item{pkg}{path to package}
}
\description{
Looks for roxygen2 function documentation in /R for author details.
Assumes that author and date are tagged via custom sections \code{@section Last updated by:}
and \code{@section Last updated date:}, respectively. To exclude a roxygen
block from this scraping, omit these section names. \cr \cr
If using a dummy documentation file, looks for \code{@name} to capture
function name, otherwise uses the actual function call. \cr \cr
Exported or internal status does not affect scraping.
}
\section{Last Updated by}{

Marie Vendettuoli
}

\section{Last updated date}{

2021-02-18
}

