% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_specification.R
\name{vt_scrape_requirements}
\alias{vt_scrape_requirements}
\title{Scrape requirement info}
\usage{
vt_scrape_requirements(
  tags = c("editor", "editDate"),
  src = ".",
  ref = vt_path(),
  dynamic_ref = NULL
)
}
\arguments{
\item{tags}{which tags to keep. defaults to editor and editDate}

\item{src}{path to package sources. defaults to current directory and passed to \code{\link{vt_scrape_tags_from}}}

\item{ref}{reference path to whre validation documentation lives. defaults
to \code{\link{vt_path}} annd passed to \code{\link{vt_scrape_tags_from}}.}

\item{dynamic_ref}{dynamic referencer object}
}
\description{
Scrape requirement info
}
