% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_section.R
\name{vt_scrape_section}
\alias{vt_scrape_section}
\title{Retrieve the value block of a custom section tagged via roxygen2}
\usage{
vt_scrape_section(tag, block)
}
\arguments{
\item{tag}{name of the section, case insensitive.}

\item{block}{character vector that holds the documentation block. \code{#' }
may be present or omitted.}
}
\value{
section value
}
\description{
Looks for the value in a custom roxygen sections. Custom
sections are named using \code{@section <NAME>:}, where colon is use to indicate
end of the name, and value starts on next line.
}
\section{Last Updated By}{

Marie Vendettuoli
}

\section{Last Updated Date}{

2021-02-18
}

\examples{
roxy_block1 <- c("@title Title1", "@param param1 definition",
    "@section Last updated date:", "2021-01-01", "@importFrom utils sessionInfo",
    "@export" )
vt_scrape_section("Last updated date:", roxy_block1)

roxy_block2 <- paste0("#' ", roxy_block1)
vt_scrape_section("Last updated date:", roxy_block2)
}
