% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_table.R
\name{vt_scrape_sig_table}
\alias{vt_scrape_sig_table}
\title{Generate a signature table for a validation report}
\usage{
vt_scrape_sig_table(usernames = NULL, pkg = ".")
}
\arguments{
\item{usernames}{list of vt_names to use when validation.yml does not exist}

\item{pkg}{Top-level directory of the package to validate}
}
\value{
A dataframe created from the validation config containing a row for
each user with the columns: role, name_and_title, signature, and date.
}
\description{
Generate a signature table for a validation report
}
