% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_val_env.R
\name{vt_scrape_val_env}
\alias{vt_scrape_val_env}
\title{Retrieve validation environment}
\usage{
vt_scrape_val_env(pkg = ".")
}
\arguments{
\item{pkg}{path to package}
}
\value{
data.frame with columns:

\itemize{
\item \code{resource} "OS", "R" or package name
\item \code{type} identifier of requirement type:
\itemize{
\item system - OS or R resource
\item package_req - from DESCRIPTION Depends/Imports of the package being validated
\item extended_req - from DESCRIPTION Suggests of the package being validated
\item session - packages in current workspace not captured via package_req/extended_req
}
\item \code{detail} OS or version details

}
}
\description{
Retrieves dependencies used in validation report.
Includes: OS, R version, packages listed in the validation package DESCRIPTION (Depends/Imports/Suggests), packages
present in current session.
}
