% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_handlers.R
\name{vt_use_news_md}
\alias{vt_use_news_md}
\title{Initiate a NEWS.md file}
\usage{
vt_use_news_md(date = NULL, version = NULL, open = interactive())
}
\arguments{
\item{date}{passed to template}

\item{version}{version to set in news file}

\item{open}{whether to open the file after}
}
\description{
Initiate a NEWS.md file
}
\note{
This is an alternative to \code{usethis::use_news_md}.
}
