% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_validation_report.R
\name{vt_use_report}
\alias{vt_use_report}
\title{Create validation report from template}
\usage{
vt_use_report(pkg = ".", template = "validation_report.Rmd")
}
\arguments{
\item{pkg}{filepath to config file}

\item{template}{what validation report template to use, passed to \code{usethis::use_template}}
}
\description{
Create validation report from template
}
