% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{vt_use_validation_config}
\alias{vt_use_validation_config}
\title{use a validation config file}
\usage{
vt_use_validation_config(
  pkg = ".",
  val_dir = "vignettes/validation",
  val_dir_o = "validation",
  username_list = list(),
  ...
)
}
\arguments{
\item{val_dir}{\link{character} where validation contents are used interactively}

\item{val_dir_o}{\link{character} which folder under \code{inst} should the contents for
validation be copied to on successful validation build.}

\item{username_list}{named list of named vectors. Name of list is username,
with named character vector containing entries for username, name, and
title to be used for documentation.}

\item{...}{These dots are for future extensions and must be empty.}

\item{path}{where to write config file}
}
\description{
validation configuration for directories and username-name-role key
}
