% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vt_username}
\alias{vt_username}
\title{Get current username}
\usage{
vt_username()
}
\value{
\verb{[character]} Username of the person that called the function
}
\description{
Wrapper for whoami::username
}
\examples{
vt_username()

}
