% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{vt_validate_installed_package}
\alias{vt_validate_installed_package}
\title{Validate an installed package}
\usage{
vt_validate_installed_package(pkg, output_directory = ".")
}
\arguments{
\item{pkg}{Name of installed package}

\item{output_directory}{Location of directory to output validation report}
}
\description{
Validate an installed package
}
