% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{vt_validated_build}
\alias{vt_validated_build}
\title{Validate a package source}
\usage{
vt_validated_build(pkg = ".", ...)
}
\arguments{
\item{pkg}{Top-level directory of the package to validate}

\item{...}{Additional argument passed to \code{devtools::build()}}
}
\description{
Validate a package source
}
