% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_markdown.R
\name{dynamic_reference_rendering}
\alias{dynamic_reference_rendering}
\title{Dynamic Reference Rendering}
\usage{
dynamic_reference_rendering(
  file,
  type = c("spec", "test_case", "test_code"),
  reference = NULL
)
}
\arguments{
\item{file}{path to the file to convert dynamic referencing to values}

\item{type}{type of file being converted; a specification ('spec'), test case ('test_case') or
test code ('test_code')}

\item{reference}{which dynamic referencer to use. When NULL, uses internal dynamic referencer.}
}
\value{
text of file with dynamic referencing evaluated
}
\description{
enable dynamic referencing by reading file and converting any dynamic references
into their values for rendering in the validation report.
}
