% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_test_code.R
\name{vt_run_test_code_file}
\alias{vt_run_test_code_file}
\alias{vt_kable_test_code_results}
\title{Evaluate the test code file}
\usage{
vt_run_test_code_file(file, test_env = new.env(), ..., ref = vt_path())

vt_kable_test_code_results(results, format = vt_render_to())
}
\arguments{
\item{file}{full path to test code file.}

\item{test_env}{environment to perform tests in}

\item{...}{argument passed to \code{knitr::kable()}}

\item{ref}{reference path to use. Defaults to vt_path()}

\item{results}{results data.frame from \code{vt_run_test_code_file()}}

\item{format}{passed to \code{knitr::kable}}
}
\value{
a kable with variables: \code{Test}, \code{Expected}, \code{Results},
\code{Pass/Fail}. Suitable for including in validation report

kableExtra object with formatting
}
\description{
Evaluate the test code file

Turn test code results data.frame into kable output
}
