% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R, R/code.R, R/req.R
\name{vt_use_test_case}
\alias{vt_use_test_case}
\alias{vt_use_test_code}
\alias{vt_use_req}
\title{Create a validation specification, test case, or test code file}
\usage{
vt_use_test_case(
  name,
  username = vt_username(),
  open = interactive(),
  pkg = "."
)

vt_use_test_code(
  name,
  username = vt_username(),
  open = interactive(),
  pkg = "."
)

vt_use_req(
  name,
  username = vt_username(),
  title = NULL,
  open = interactive(),
  pkg = "."
)
}
\arguments{
\item{name}{The name/path of the validation item. These can be named with your
file system separator and will be organized as a directory structure. Items
are located at \verb{./inst/validation/<ItemType>/\{name\}}.}

\item{username}{The username to insert into the specification as the author.}

\item{open}{Should the newly made file be opened for editing.}

\item{pkg}{Top-level of package}

\item{title}{Title for the specification. defaults to be the base name passed
sans file paths or extensions.}
}
\value{
Path to the newly created validation item file, invisibly.
}
\description{
Create a validation specification, test case, or test code file
}
\examples{
withr::with_tempdir({

vt_create_package(package_dir)
# Create spec at the cases top level `inst/validation/cases/spec1`
vt_use_test_case("spec1", pkg = package_dir, open = FALSE)
# Create spec at `inst/validation/cases/regTests/Update2/spec2`
vt_use_test_case("regTests/Update2/spec2", pkg = package_dir, open = FALSE)

})
}
