% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R, R/code.R, R/spec.R
\name{vt_use_test_case}
\alias{vt_use_test_case}
\alias{vt_use_test_code}
\alias{vt_use_spec}
\title{Create a validation specification, test case, or test code file}
\usage{
vt_use_test_case(
  name,
  username = vt_username(),
  open = interactive(),
  pkg = "."
)

vt_use_test_code(
  name,
  username = vt_username(),
  open = interactive(),
  pkg = "."
)

vt_use_spec(
  name,
  username = vt_username(),
  title = NULL,
  open = interactive(),
  pkg = "."
)
}
\arguments{
\item{name}{The Name of the new specification. specs can be named with your
file system separator and will be organized as a directory structure. Specs
are located at \code{./inst/validation/specification/{spec}}.}

\item{username}{the username to insert into the specification as the author.}

\item{open}{should the newly make specification open for editing.}

\item{pkg}{Top-level of package}

\item{title}{title for the specification. defaults to be the base name passed sans
file paths or extensions.}
}
\value{
Path to the newly created specification file, invisibly.
}
\description{
Create a validation specification, test case, or test code file
}
\examples{
\dontrun{
# TODO: error during devtools::check
package_dir <- tempdir()

vt_use_validation(package_dir)
# Create spec at the cases top level `inst/validation/cases/spec1`
vt_use_test_case("spec1", package_dir)
# Create spec at `inst/validation/cases/regTests/Update2/spec2`
vt_use_test_case("regTests/Update2/spec2", package_dir)
}
}
