% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R, R/code.R, R/req.R
\name{vt_use_test_case}
\alias{vt_use_test_case}
\alias{vt_use_test_code}
\alias{vt_use_req}
\title{Create a validation requirement, test case, or test code file}
\usage{
vt_use_test_case(
  name,
  username = vt_username(),
  title = NULL,
  open = interactive()
)

vt_use_test_code(name, username = vt_username(), open = interactive())

vt_use_req(name, username = vt_username(), title = NULL, open = interactive())
}
\arguments{
\item{name}{The name/path of the validation item. These can be named with your
file system separator and will be organized as a directory structure. Items
are located at \verb{./inst/validation/<ItemType>/\{name\}}.}

\item{username}{The username to insert into the validation item as the author.}

\item{title}{Title for the requirement defaults to be the base name passed
sans file paths or extensions.}

\item{open}{Should the newly made file be opened for editing.}
}
\value{
Path to the newly created validation item file, invisibly.
}
\description{
Create a validation requirement, test case, or test code file
}
\examples{
withr::with_tempdir({
vt_create_package("example.package")
setwd("example.package")
vt_add_user_to_config(
  username = whoami::username(),
  name = "Sample Name",
  title = "Sample",
  role = "example"
 )
# Create req at the cases top level `inst/validation/cases/case1`
vt_use_test_case("case1", open = FALSE)
# Create req at `inst/validation/cases/regTests/Update2/case2`
vt_use_test_case("regTests/Update2/case2", open = FALSE)

})
}
