% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{vt_validate_source}
\alias{vt_validate_source}
\alias{vt_validate_build}
\alias{vt_validate_installed_package}
\title{Validate a package}
\usage{
vt_validate_source(pkg = ".", open = interactive())

vt_validate_build(pkg = ".", ...)

vt_validate_installed_package(
  package,
  output_directory = ".",
  open = interactive()
)
}
\arguments{
\item{pkg}{Top-level directory of the package to validate}

\item{open}{should the validation report be opened after it is built?}

\item{...}{Additional argument passed to \code{devtools::build()}}

\item{package}{installed package name}

\item{output_directory}{Location of directory to output validation report}
}
\value{
path to either the validation report or the bundled package
}
\description{
\code{vt_validate_source} runs the validation on the current source, temporarily
installing the source to properly evaluate the report. \code{vt_validate_build()}
runs the same step, then compiles a bundle that includes th the validation
report and any other contents that are required for validation. Finally,
\code{vt_validate_installed_package} run rerun the validation report for packages
that were built and then installed using the\code{vt_validate_build()}.
}
