% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_validation_report.R
\name{vt_path}
\alias{vt_path}
\title{Use to set dynamic file paths in a validation.}
\usage{
vt_path(...)
}
\arguments{
\item{...}{\verb{[character]}\cr
Path components below the validation folder, can be empty.
Each argument should be a string containing one or more
path components separated by a forward slash \code{"/"}.}
}
\description{
Use to set dynamic file paths in a validation.
}
\examples{
vt_path()
\dontrun{
vt_path("some", "reqs", "req01.md")
vt_path("some/reqs/req01.md")
}


}
