% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_validation_report.R
\name{vt_path}
\alias{vt_path}
\title{Use to set dynamic file paths in a validation.}
\usage{
vt_path(..., pkg = ".")
}
\arguments{
\item{...}{\verb{[character]}\cr
Path components below the validation folder, can be empty.
Each argument should be a string containing one or more
path components separated by a forward slash \code{"/"}.}

\item{pkg}{path to base directory of package}
}
\description{
vt_path() allows access of files relative to the working directory identified
in the config file. It is also required to be used in the validation report
for cases where validation of installed packages is intended as it will
shift access to the correct location for the installed package for access.
}
\examples{
withr::with_tempdir({callr::r(function(){

 valtools::vt_use_validation_config()
 valtools::vt_use_validation()

 valtools::vt_path()
 valtools::vt_path("some", "reqs", "req01.md")
 valtools::vt_path("some/reqs/req01.md")

})})

}
