% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_test_code.R
\name{vt_run_test_code_file}
\alias{vt_run_test_code_file}
\title{Evaluate the test code file}
\usage{
vt_run_test_code_file(file, test_env = new.env(), ..., ref = vt_path())
}
\arguments{
\item{test_env}{environment to perform tests in}

\item{...}{argument passed to \code{knitr::kable()}}

\item{ref}{reference path to use. Defaults to vt_path()}

\item{file.}{full path to test code file.}
}
\value{
a kable with variables: \code{Test}, \code{Expected}, \code{Results},
\code{Pass/Fail}. Suitable for including in validation report
}
\description{
Evaluate the test code file
}
