% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_handlers.R
\name{vt_scrape_news}
\alias{vt_scrape_news}
\title{Scrape news from a package}
\usage{
vt_scrape_news(pkg = rprojroot::find_root(criterion = has_file("DESCRIPTION")))
}
\arguments{
\item{pkg}{installed package}
}
\value{
data.frame with variables \code{version}, \code{effective_date}, \code{description}
}
\description{
Scrape news from a package
}
\note{
Uses the function \link{utils::news} to extract version and date. Description
are news items that start with \code{[validation]}. If no date is provided, will
look for NEWs entries starting with \code{[version date]}
}
