% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_handlers.R
\name{vt_use_news_md}
\alias{vt_use_news_md}
\title{Initiate a NEWS.md file}
\usage{
vt_use_news_md(date = NULL)
}
\arguments{
\item{date}{passed to template}
}
\description{
Initiate a NEWS.md file
}
\note{
This is an alternative to \code{usethis::use_news_md}.
}
