% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_handlers.R
\name{vt_use_news_md}
\alias{vt_use_news_md}
\title{Initiate a NEWS.md file}
\usage{
vt_use_news_md(date = NULL, open = rlang::is_interactive())
}
\arguments{
\item{open}{Open the newly created file for editing?  Happens in RStudio, if applicable, or via utils::file.edit() otherwise.}

\item{data}{passed to \code{usethis::use_template}}
}
\description{
Initiate a NEWS.md file
}
\note{
This is an alternative to \code{usethis::use_news_md}.
}
