% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_validation_report.R
\name{vt_use_report}
\alias{vt_use_report}
\title{Create validation report from template}
\usage{
vt_use_report(
  pkg_name = desc::desc_get_field("Package"),
  template = "validation",
  open = FALSE
)
}
\arguments{
\item{pkg_name}{name of package}

\item{template}{what validation report template from {valtools} to use,
one of "validation" (default) or "requirements" (forthcoming)}
}
\description{
Create validation report from template
}
