% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_validation_report.R
\name{vt_use_report}
\alias{vt_use_report}
\title{Create validation report from template}
\usage{
vt_use_report(
  pkg_name = desc::desc_get_field("Package"),
  template = "validation_report.Rmd",
  open = FALSE
)
}
\arguments{
\item{pkg_name}{name of package}

\item{template}{what validation report template from {valtools} to use,
passed to \code{usethis::use_template}}

\item{open}{boolean whether to open the validation report, passed to \code{usetthis::use_template}}
}
\description{
Create validation report from template
}
