% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_validation_report.R
\name{vt_use_report}
\alias{vt_use_report}
\title{Create validation report from template}
\usage{
vt_use_report(
  pkg_name = NULL,
  template = "validation",
  dynamic_referencing = FALSE,
  open = is_interactive()
)
}
\arguments{
\item{pkg_name}{name of package}

\item{template}{what validation report template from {valtools} to use,
one of "validation" (default) or "requirements"}

\item{dynamic_referencing}{Should dynamic referencing be enabled by default. Boolean defaults to FALSE.}

\item{open}{boolean to open the validation report for further editing}
}
\description{
Create validation report from template
}
