% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vt_username}
\alias{vt_username}
\title{Get current username}
\usage{
vt_username(pkg = ".")
}
\arguments{
\item{pkg}{path to base directory of working project}
}
\value{
\verb{[character]} Username of the person that called the function
}
\description{
Wrapper for whoami::username
}
\examples{
\dontrun{
vt_use_validation_config(
    usename_list = list(vt_user(
      username = whoami::username(),
      name = "test",
      title = "title",
      role = "role")))
vt_username()
}
}
