% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vt_username}
\alias{vt_username}
\title{Get current username}
\usage{
vt_username()
}
\description{
Wrapper for whoami::username
}
\details{
@returns \verb{[character]} Username of the person that called the function
}
\examples{
withr::with_tempdir({
vt_use_validation_(
    username_list = list(vt_user(
      username = whoami::username(),
      name = "test",
      title = "title",
      role = "role")))
vt_username()
})

}
